books_database/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
watchdog==6.0.0
docker==7.1.0